+++ src/theme/prose/field-group.ts
@@ -1,0 +1,9 @@
+/**
+ * ProseFieldGroup
+ * ---
+ * @todo fix color
+ */
+
+export default {
+  base: 'my-[20px] divide-y divide-(--ui-color-design-tinted-na-stroke) *:not-last:pb-[20px]'
+}